requirements.txt
alpha_shapes==1.1.0
alphashape==1.3.1
einops==0.8.0
matplotlib==3.7.1
meshio==5.3.5
multipledispatch==1.0.0
networkx==3.2.1
numpy==1.26.4
open3d==0.17.0
Pillow==11.1.0
psutil==5.9.0
scikit_learn==1.3.0
tqdm==4.65.0
trimesh==4.0.0
scipy
torchinfo
ftfy 
regex
mmengine
git+https://github.com/openai/CLIP.git